+++ JenkinsTrayTracker.Tests/JenkinsTrayTracker.Tests.csproj
