--- setup/HudsonTrayTracker.WixSetup.sln
@@ -1,28 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "JenkinsTrayTracker.WixSetup", "JenkinsTrayTracker.WixSetup\JenkinsTrayTracker.WixSetup.wixproj", "{4839B061-A1DF-4F9F-852E-95F534D949BB}"
-	ProjectSection(ProjectDependencies) = postProject
-		{95AA7347-A21A-4F09-80DF-D9797B0FA997} = {95AA7347-A21A-4F09-80DF-D9797B0FA997}
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JenkinsTrayTracker.Setup.CustomActions", "JenkinsTrayTracker.Setup.CustomActions\JenkinsTrayTracker.Setup.CustomActions.csproj", "{95AA7347-A21A-4F09-80DF-D9797B0FA997}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4839B061-A1DF-4F9F-852E-95F534D949BB}.Debug|x86.ActiveCfg = Debug|x86
-		{4839B061-A1DF-4F9F-852E-95F534D949BB}.Debug|x86.Build.0 = Debug|x86
-		{4839B061-A1DF-4F9F-852E-95F534D949BB}.Release|x86.ActiveCfg = Release|x86
-		{4839B061-A1DF-4F9F-852E-95F534D949BB}.Release|x86.Build.0 = Release|x86
-		{95AA7347-A21A-4F09-80DF-D9797B0FA997}.Debug|x86.ActiveCfg = Debug|x86
-		{95AA7347-A21A-4F09-80DF-D9797B0FA997}.Debug|x86.Build.0 = Debug|x86
-		{95AA7347-A21A-4F09-80DF-D9797B0FA997}.Release|x86.ActiveCfg = Release|x86
-		{95AA7347-A21A-4F09-80DF-D9797B0FA997}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
