--- setup/HudsonTrayTracker.WixSetup/HudsonTrayTracker.WixSetup.wixproj
@@ -1,61 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">x86</Platform>
-    <ProjectGuid>{4839b061-a1df-4f9f-852e-95f534d949bb}</ProjectGuid>
-    <SchemaVersion>2.0</SchemaVersion>
-    <OutputName>JenkinsTray</OutputName>
-    <OutputType>Package</OutputType>
-    <WixTargetsPath Condition=" '$(WixTargetsPath)' == '' AND '$(MSBuildExtensionsPath32)' != '' ">$(MSBuildExtensionsPath32)\Microsoft\WiX\v3.x\Wix.targets</WixTargetsPath>
-    <WixTargetsPath Condition=" '$(WixTargetsPath)' == '' ">$(MSBuildExtensionsPath)\Microsoft\WiX\v3.x\Wix.targets</WixTargetsPath>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
-    <OutputPath>bin\$(Configuration)\</OutputPath>
-    <IntermediateOutputPath>obj\$(Configuration)\</IntermediateOutputPath>
-    <DefineConstants>Debug</DefineConstants>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
-    <OutputPath>bin\$(Configuration)\</OutputPath>
-    <IntermediateOutputPath>obj\$(Configuration)\</IntermediateOutputPath>
-    <DefineConstants></DefineConstants>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="JenkinsTrayTracker.wxs" />
-  </ItemGroup>
-  <ItemGroup>
-    <WixExtension Include="WixUtilExtension">
-      <HintPath>$(WixExtDir)\WixUtilExtension.dll</HintPath>
-      <Name>WixUtilExtension</Name>
-    </WixExtension>
-    <WixExtension Include="WixUIExtension">
-      <HintPath>$(WixExtDir)\WixUIExtension.dll</HintPath>
-      <Name>WixUIExtension</Name>
-    </WixExtension>
-  </ItemGroup>
-  <Import Project="$(WixTargetsPath)" />
-
-  <UsingTask AssemblyFile="..\JenkinsTrayTracker.Setup.CustomActions\bin\$(Platform)\$(Configuration)\JenkinsTray.Setup.CustomActions.dll" TaskName="JenkinsTray.Setup.CustomActions.GetAssemblyFileVersion" />
-  <Target Name="BeforeBuild">
-    <!-- Get the programs assembly version from the .exe file -->
-    <GetAssemblyFileVersion strFilePathAssemblyInfo="..\..\common\VersionInfo.cs">
-      <Output TaskParameter="strAssemblyFileVersion" PropertyName="strAssemblyFileVersion" />
-    </GetAssemblyFileVersion>
-    <!-- Store the assembly version number in ProductVersion preprocessor variable -->
-    <CreateProperty Value="$(DefineConstants);ProductVersion=$(strAssemblyFileVersion)">
-      <Output TaskParameter="Value" PropertyName="DefineConstants" />
-    </CreateProperty>
-    <!-- Name the .msi file after the solution platform and assembly version e.g TestService-x86-1.4.0.0.msi -->
-    <CreateProperty Value="$(OutputName)_v$(strAssemblyFileVersion)">
-      <Output TaskParameter="Value" PropertyName="TargetName" />
-    </CreateProperty>
-    <!-- Name the .wixpdb file after the solution platform and assembly version e.g TestService-x86-1.4.0.0.msi -->
-    <CreateProperty Value="$(TargetName)$(TargetPdbExt)">
-      <Output TaskParameter="Value" PropertyName="TargetPdbName" />
-    </CreateProperty>
-  </Target>
-
-  <!-- Opens the output directory after build -->
-  <Target Name="AfterBuild">
-  </Target>
-</Project>
