+++ setup/JenkinsTrayTracker.WixSetup/src/InstallUtilConfig.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<configuration>
+<startup>
+<supportedRuntime version="v2.0.50727"/>
+</startup>
+</configuration>
